Add autotune support.
Note that the Group-Size input must have consistent data. The autotuner has been updated to enable that.
