+++ xla/backends/gpu/transforms/gemm_rewriter_test.cc
@@ -56,700 +56,6 @@ namespace {
 
 namespace m = ::xla::match;
 
-using GemmRewriteTest = GemmRewriteTestBase;
-
-class GroupedGemmRewriteTest : public GemmRewriteTest {
- public:
-  DebugOptions GetDebugOptionsForTest() const override {
-    DebugOptions debug_options = GemmRewriteTest::GetDebugOptionsForTest();
-    debug_options.set_xla_gpu_experimental_use_ragged_dot_grouped_gemm(true);
-    debug_options.set_xla_gpu_enable_cublaslt(true);
-    return debug_options;
-  }
-};
-
-TEST_F(GroupedGemmRewriteTest, CheckCustomCallTargetGroupedGemm) {
-  if (SkipGpuBlasLtTest()) {
-    GTEST_SKIP() << "BlasLt is not supported on this GPU architecture";
-  }
-  const char* hlo_text = R"(
-HloModule GroupedGemm
-
-ENTRY AddRaggedDotsFunc {
-    p0 = bf16[64,9]{1,0} parameter(0)
-    p1 = bf16[2,9,8]{2,1,0} parameter(1)
-    p2 = s32[2] constant({16, 48})
-    ROOT ragged-dot = bf16[64,8]{1,0} ragged-dot(p0, p1, p2),
-                      lhs_contracting_dims={1}, rhs_contracting_dims={1},
-                      lhs_ragged_dims={0}, rhs_group_dims={0}
-}
-)";
-  MatchOptimizedHlo(hlo_text,
-                    R"(
-                    ; CHECK: custom_call_target="__cublas$lt$groupedMatmul", 
-                    ; CHECK-SAME: backend_config={"operation_queue_id":"0","wait_on_operation_queues":[],
-                    ; CHECK-SAME: "force_earliest_schedule":false,"reification_cost":[],
-                    ; CHECK-SAME: "device_type":"DEVICE_TYPE_INVALID",
-                    ; CHECK-SAME: "grouped_gemm_backend_config":{
-                    ; CHECK-SAME: "gemm_backend_config":{"alpha_real":1,"beta":0,
-                    ; CHECK-SAME: "dot_dimension_numbers":{"lhs_contracting_dimensions":["1"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["1"],"lhs_batch_dimensions":[],
-                    ; CHECK-SAME: "rhs_batch_dimensions":[]},"alpha_imag":0,"epilogue":"DEFAULT",
-                    ; CHECK-SAME: "grad_x":false,"grad_y":false,
-                    ; CHECK-SAME: "damax_output":false,"autotune_workspace_size":"0"},
-                    ; CHECK-SAME: "ragged_dot_dimension_numbers":{"dot_dimension_numbers":{
-                    ; CHECK-SAME: "lhs_contracting_dimensions":["1"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["1"],
-                    ; CHECK-SAME: "lhs_batch_dimensions":[],
-                    ; CHECK-SAME: "rhs_batch_dimensions":[]},
-                    ; CHECK-SAME: "lhs_ragged_dimensions":["0"],
-                    ; CHECK-SAME: "rhs_group_dimensions":["0"]}}})");
-  EXPECT_TRUE(RunAndCompare(hlo_text, ErrorSpec{1e-4, 1e-5}));
-}
-
-TEST_F(GroupedGemmRewriteTest,
-       CheckCustomCallTargetGroupedGemmLargeGroupCount600) {
-  if (SkipGpuBlasLtTest()) {
-    GTEST_SKIP() << "BlasLt is not supported on this GPU architecture";
-  }
-  // Test with 600 groups (> 256) to test single batch processing in grid-stride
-  // loop Sum of group sizes: 600 * 4 = 2400 (matches p0 first dim)
-  const char* hlo_text = R"(
-HloModule GroupedGemmLarge600
-
-ENTRY AddRaggedDotsFunc {
-    p0 = bf16[2400,9]{1,0} parameter(0)
-    p1 = bf16[600,9,8]{2,1,0} parameter(1)
-    scalar = s32[] constant(4)
-    p2 = s32[600] broadcast(scalar), dimensions={}
-    ROOT ragged-dot = bf16[2400,8]{1,0} ragged-dot(p0, p1, p2),
-                      lhs_contracting_dims={1}, rhs_contracting_dims={1},
-                      lhs_ragged_dims={0}, rhs_group_dims={0}
-}
-)";
-  MatchOptimizedHlo(hlo_text,
-                    R"(
-                    ; CHECK: custom_call_target="__cublas$lt$groupedMatmul")");
-  EXPECT_TRUE(RunAndCompare(hlo_text, ErrorSpec{1e-2, 1e-2}));
-}
-
-TEST_F(GroupedGemmRewriteTest, CheckCustomCallTargetGroupedGemmMulipleGroups) {
-  if (SkipGpuBlasLtTest()) {
-    GTEST_SKIP() << "BlasLt is not supported on this GPU architecture";
-  }
-  const char* hlo_text = R"(
-HloModule GroupedGemm
-
-ENTRY AddRaggedDotsFunc {
-    p0 = bf16[64,9]{1,0} parameter(0)
-    p1 = bf16[4,9,8]{2,1,0} parameter(1)
-    p2 = s64[4] constant({16, 8, 24, 16})
-    ROOT ragged-dot = bf16[64,8]{1,0} ragged-dot(p0, p1, p2),
-                      lhs_contracting_dims={1}, rhs_contracting_dims={1},
-                      lhs_ragged_dims={0}, rhs_group_dims={0}
-}
-)";
-  MatchOptimizedHlo(hlo_text,
-                    R"(
-                    ; CHECK: custom_call_target="__cublas$lt$groupedMatmul", 
-                    ; CHECK-SAME: backend_config={"operation_queue_id":"0","wait_on_operation_queues":[],
-                    ; CHECK-SAME: "force_earliest_schedule":false,"reification_cost":[],
-                    ; CHECK-SAME: "device_type":"DEVICE_TYPE_INVALID",
-                    ; CHECK-SAME: "grouped_gemm_backend_config":{
-                    ; CHECK-SAME: "gemm_backend_config":{"alpha_real":1,"beta":0,
-                    ; CHECK-SAME: "dot_dimension_numbers":{"lhs_contracting_dimensions":["1"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["1"],"lhs_batch_dimensions":[],
-                    ; CHECK-SAME: "rhs_batch_dimensions":[]},"alpha_imag":0,"epilogue":"DEFAULT",
-                    ; CHECK-SAME: "grad_x":false,"grad_y":false,
-                    ; CHECK-SAME: "damax_output":false,"autotune_workspace_size":"0"},
-                    ; CHECK-SAME: "ragged_dot_dimension_numbers":{"dot_dimension_numbers":{
-                    ; CHECK-SAME: "lhs_contracting_dimensions":["1"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["1"],
-                    ; CHECK-SAME: "lhs_batch_dimensions":[],
-                    ; CHECK-SAME: "rhs_batch_dimensions":[]},
-                    ; CHECK-SAME: "lhs_ragged_dimensions":["0"],
-                    ; CHECK-SAME: "rhs_group_dimensions":["0"]}}})");
-  EXPECT_TRUE(RunAndCompare(hlo_text, ErrorSpec{1e-4, 1e-5}));
-}
-
-TEST_F(GroupedGemmRewriteTest,
-       CheckCustomCallTargetGroupedGemmMulipleGroupsOutputColumnMajor) {
-  if (SkipGpuBlasLtTest()) {
-    GTEST_SKIP() << "BlasLt is not supported on this GPU architecture";
-  }
-  const char* hlo_text = R"(
-HloModule GroupedGemm
-
-ENTRY AddRaggedDotsFunc {
-    p0 = bf16[64,9]{1,0} parameter(0)
-    p1 = bf16[4,9,8]{2,1,0} parameter(1)
-    p2 = s64[4] constant({16, 8, 24, 16})
-    ROOT ragged-dot = bf16[64,8]{0,1} ragged-dot(p0, p1, p2),
-                      lhs_contracting_dims={1}, rhs_contracting_dims={1},
-                      lhs_ragged_dims={0}, rhs_group_dims={0}
-}
-)";
-  MatchOptimizedHlo(hlo_text,
-                    R"(
-                    ; CHECK: custom_call_target="__cublas$lt$groupedMatmul", 
-                    ; CHECK-SAME: backend_config={"operation_queue_id":"0","wait_on_operation_queues":[],
-                    ; CHECK-SAME: "force_earliest_schedule":false,"reification_cost":[],
-                    ; CHECK-SAME: "device_type":"DEVICE_TYPE_INVALID",
-                    ; CHECK-SAME: "grouped_gemm_backend_config":{
-                    ; CHECK-SAME: "gemm_backend_config":{"alpha_real":1,"beta":0,
-                    ; CHECK-SAME: "dot_dimension_numbers":{"lhs_contracting_dimensions":["1"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["1"],"lhs_batch_dimensions":[],
-                    ; CHECK-SAME: "rhs_batch_dimensions":[]},"alpha_imag":0,"epilogue":"DEFAULT",
-                    ; CHECK-SAME: "grad_x":false,"grad_y":false,
-                    ; CHECK-SAME: "damax_output":false,"autotune_workspace_size":"0"},
-                    ; CHECK-SAME: "ragged_dot_dimension_numbers":{"dot_dimension_numbers":{
-                    ; CHECK-SAME: "lhs_contracting_dimensions":["1"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["1"],
-                    ; CHECK-SAME: "lhs_batch_dimensions":[],
-                    ; CHECK-SAME: "rhs_batch_dimensions":[]},
-                    ; CHECK-SAME: "lhs_ragged_dimensions":["0"],
-                    ; CHECK-SAME: "rhs_group_dimensions":["0"]}}})");
-  EXPECT_TRUE(RunAndCompare(hlo_text, ErrorSpec{1e-4, 1e-5}));
-}
-
-TEST_F(GroupedGemmRewriteTest,
-       CheckCustomCallTargetGroupedGemmNonContractingWithBatchDim) {
-  if (SkipGpuBlasLtTest()) {
-    GTEST_SKIP() << "BlasLt is not supported on this GPU architecture";
-  }
-  const char* hlo_text = R"(
-HloModule GroupedGemm
-
-ENTRY AddRaggedDotsFunc {
-    p0 = bf16[3,16,9]{2,1,0} parameter(0)
-    p1 = bf16[3,2,9,8]{3,2,1,0} parameter(1)
-    p2 = s64[3, 2] constant({{4, 12}, {4, 12}, {4, 12}})
-    ROOT ragged-dot = bf16[3,16,8]{2,1,0} ragged-dot(p0, p1, p2),
-                      lhs_contracting_dims={2}, rhs_contracting_dims={2},
-                      lhs_batch_dims={0}, rhs_batch_dims={0},
-                      lhs_ragged_dims={1}, rhs_group_dims={1}
-}
-)";
-  MatchOptimizedHlo(hlo_text,
-                    R"(
-                    ; CHECK: custom_call_target="__cublas$lt$groupedMatmul", 
-                    ; CHECK-SAME: backend_config={"operation_queue_id":"0","wait_on_operation_queues":[],
-                    ; CHECK-SAME: "force_earliest_schedule":false,"reification_cost":[],
-                    ; CHECK-SAME: "device_type":"DEVICE_TYPE_INVALID",
-                    ; CHECK-SAME: "grouped_gemm_backend_config":{
-                    ; CHECK-SAME: "gemm_backend_config":{"alpha_real":1,"beta":0,
-                    ; CHECK-SAME: "dot_dimension_numbers":{"lhs_contracting_dimensions":["2"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["2"],"lhs_batch_dimensions":["0"],
-                    ; CHECK-SAME: "rhs_batch_dimensions":["0"]},"alpha_imag":0,"epilogue":"DEFAULT",
-                    ; CHECK-SAME: "grad_x":false,"grad_y":false,
-                    ; CHECK-SAME: "damax_output":false,"autotune_workspace_size":"0"},
-                    ; CHECK-SAME: "ragged_dot_dimension_numbers":{"dot_dimension_numbers":{
-                    ; CHECK-SAME: "lhs_contracting_dimensions":["2"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["2"],
-                    ; CHECK-SAME: "lhs_batch_dimensions":["0"],
-                    ; CHECK-SAME: "rhs_batch_dimensions":["0"]},
-                    ; CHECK-SAME: "lhs_ragged_dimensions":["1"],
-                    ; CHECK-SAME: "rhs_group_dimensions":["1"]}}})");
-  EXPECT_TRUE(RunAndCompare(hlo_text, ErrorSpec{1e-4, 1e-5}));
-}
-
-TEST_F(GroupedGemmRewriteTest,
-       CheckCustomCallTargetGroupedGemmRaggedDimInContractingDim) {
-  if (SkipGpuBlasLtTest()) {
-    GTEST_SKIP() << "BlasLt is not supported on this GPU architecture";
-  }
-  const char* hlo_text = R"(
-HloModule GroupedGemm
-
-ENTRY AddRaggedDotsFunc {
-    p0 = bf16[64,9]{1,0} parameter(0)
-    p1 = bf16[9,8]{1,0} parameter(1)
-    p2 = s64[2] constant({4, 5})
-    ROOT ragged-dot = bf16[2,64,8]{2,1,0} ragged-dot(p0, p1, p2),
-                      lhs_contracting_dims={1}, rhs_contracting_dims={0},
-                      lhs_ragged_dims={1}
-}
-)";
-  MatchOptimizedHlo(hlo_text,
-                    R"(
-                    ; CHECK: custom_call_target="__cublas$lt$groupedMatmul", 
-                    ; CHECK-SAME: backend_config={"operation_queue_id":"0","wait_on_operation_queues":[],
-                    ; CHECK-SAME: "force_earliest_schedule":false,"reification_cost":[],
-                    ; CHECK-SAME: "device_type":"DEVICE_TYPE_INVALID",
-                    ; CHECK-SAME: "grouped_gemm_backend_config":{
-                    ; CHECK-SAME: "gemm_backend_config":{"alpha_real":1,"beta":0,
-                    ; CHECK-SAME: "dot_dimension_numbers":{"lhs_contracting_dimensions":["1"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["0"],"lhs_batch_dimensions":[],
-                    ; CHECK-SAME: "rhs_batch_dimensions":[]},"alpha_imag":0,"epilogue":"DEFAULT",
-                    ; CHECK-SAME: "grad_x":false,"grad_y":false,
-                    ; CHECK-SAME: "damax_output":false,"autotune_workspace_size":"0"},
-                    ; CHECK-SAME: "ragged_dot_dimension_numbers":{"dot_dimension_numbers":{
-                    ; CHECK-SAME: "lhs_contracting_dimensions":["1"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["0"],
-                    ; CHECK-SAME: "lhs_batch_dimensions":[],
-                    ; CHECK-SAME: "rhs_batch_dimensions":[]},
-                    ; CHECK-SAME: "lhs_ragged_dimensions":["1"],
-                    ; CHECK-SAME: "rhs_group_dimensions":[]}}})");
-  EXPECT_TRUE(RunAndCompare(hlo_text, ErrorSpec{1e-4, 1e-5}));
-}
-
-TEST_F(
-    GroupedGemmRewriteTest,
-    CheckCustomCallTargetGroupedGemmRaggedDimInContractingDimMultipleGroups) {
-  if (SkipGpuBlasLtTest()) {
-    GTEST_SKIP() << "BlasLt is not supported on this GPU architecture";
-  }
-  const char* hlo_text = R"(
-HloModule GroupedGemm
-
-ENTRY AddRaggedDotsFunc {
-    p0 = bf16[64,16]{1,0} parameter(0)
-    p1 = bf16[16,8]{1,0} parameter(1)
-    p2 = s64[4] constant({4, 5, 3, 4})
-    ROOT ragged-dot = bf16[4,64,8]{2,1,0} ragged-dot(p0, p1, p2),
-                      lhs_contracting_dims={1}, rhs_contracting_dims={0},
-                      lhs_ragged_dims={1}
-}
-)";
-  MatchOptimizedHlo(hlo_text,
-                    R"(
-                    ; CHECK: custom_call_target="__cublas$lt$groupedMatmul", 
-                    ; CHECK-SAME: backend_config={"operation_queue_id":"0","wait_on_operation_queues":[],
-                    ; CHECK-SAME: "force_earliest_schedule":false,"reification_cost":[],
-                    ; CHECK-SAME: "device_type":"DEVICE_TYPE_INVALID",
-                    ; CHECK-SAME: "grouped_gemm_backend_config":{
-                    ; CHECK-SAME: "gemm_backend_config":{"alpha_real":1,"beta":0,
-                    ; CHECK-SAME: "dot_dimension_numbers":{"lhs_contracting_dimensions":["1"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["0"],"lhs_batch_dimensions":[],
-                    ; CHECK-SAME: "rhs_batch_dimensions":[]},"alpha_imag":0,"epilogue":"DEFAULT",
-                    ; CHECK-SAME: "grad_x":false,"grad_y":false,
-                    ; CHECK-SAME: "damax_output":false,"autotune_workspace_size":"0"},
-                    ; CHECK-SAME: "ragged_dot_dimension_numbers":{"dot_dimension_numbers":{
-                    ; CHECK-SAME: "lhs_contracting_dimensions":["1"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["0"],
-                    ; CHECK-SAME: "lhs_batch_dimensions":[],
-                    ; CHECK-SAME: "rhs_batch_dimensions":[]},
-                    ; CHECK-SAME: "lhs_ragged_dimensions":["1"],
-                    ; CHECK-SAME: "rhs_group_dimensions":[]}}})");
-  EXPECT_TRUE(RunAndCompare(hlo_text, ErrorSpec{1e-4, 1e-5}));
-}
-
-TEST_F(
-    GroupedGemmRewriteTest,
-    CheckCustomCallTargetGroupedGemmRaggedDimInContractingDimMultipleGroupsOutputColumnMajor) {
-  if (SkipGpuBlasLtTest()) {
-    GTEST_SKIP() << "BlasLt is not supported on this GPU architecture";
-  }
-  const char* hlo_text = R"(
-HloModule GroupedGemm
-
-ENTRY AddRaggedDotsFunc {
-    p0 = bf16[64,16]{1,0} parameter(0)
-    p1 = bf16[16,8]{1,0} parameter(1)
-    p2 = s64[4] constant({4, 5, 3, 4})
-    ROOT ragged-dot = bf16[4,64,8]{1,2,0} ragged-dot(p0, p1, p2),
-                      lhs_contracting_dims={1}, rhs_contracting_dims={0},
-                      lhs_ragged_dims={1}
-}
-)";
-  MatchOptimizedHlo(hlo_text,
-                    R"(
-                    ; CHECK: custom_call_target="__cublas$lt$groupedMatmul", 
-                    ; CHECK-SAME: backend_config={"operation_queue_id":"0","wait_on_operation_queues":[],
-                    ; CHECK-SAME: "force_earliest_schedule":false,"reification_cost":[],
-                    ; CHECK-SAME: "device_type":"DEVICE_TYPE_INVALID",
-                    ; CHECK-SAME: "grouped_gemm_backend_config":{
-                    ; CHECK-SAME: "gemm_backend_config":{"alpha_real":1,"beta":0,
-                    ; CHECK-SAME: "dot_dimension_numbers":{"lhs_contracting_dimensions":["1"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["0"],"lhs_batch_dimensions":[],
-                    ; CHECK-SAME: "rhs_batch_dimensions":[]},"alpha_imag":0,"epilogue":"DEFAULT",
-                    ; CHECK-SAME: "grad_x":false,"grad_y":false,
-                    ; CHECK-SAME: "damax_output":false,"autotune_workspace_size":"0"},
-                    ; CHECK-SAME: "ragged_dot_dimension_numbers":{"dot_dimension_numbers":{
-                    ; CHECK-SAME: "lhs_contracting_dimensions":["1"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["0"],
-                    ; CHECK-SAME: "lhs_batch_dimensions":[],
-                    ; CHECK-SAME: "rhs_batch_dimensions":[]},
-                    ; CHECK-SAME: "lhs_ragged_dimensions":["1"],
-                    ; CHECK-SAME: "rhs_group_dimensions":[]}}})");
-  EXPECT_TRUE(RunAndCompare(hlo_text, ErrorSpec{1e-4, 1e-5}));
-}
-
-TEST_F(
-    GroupedGemmRewriteTest,
-    CheckCustomCallTargetGroupedGemmRaggedDimInContractingDimLargeGroupCount600) {
-  if (SkipGpuBlasLtTest()) {
-    GTEST_SKIP() << "BlasLt is not supported on this GPU architecture";
-  }
-  // Test with 600 groups (> 256) for RaggedDimInContractingDim
-  // Sum of group sizes: 600 * 4 = 2400 (matches p1 first dim for contracting)
-  const char* hlo_text = R"(
-HloModule GroupedGemmRaggedContractingLarge600
-
-ENTRY AddRaggedDotsFunc {
-    p0 = bf16[64,2400]{1,0} parameter(0)
-    p1 = bf16[2400,8]{1,0} parameter(1)
-    scalar = s64[] constant(4)
-    p2 = s64[600] broadcast(scalar), dimensions={}
-    ROOT ragged-dot = bf16[600,64,8]{2,1,0} ragged-dot(p0, p1, p2),
-                      lhs_contracting_dims={1}, rhs_contracting_dims={0},
-                      lhs_ragged_dims={1}
-}
-)";
-  MatchOptimizedHlo(hlo_text,
-                    R"(
-                    ; CHECK: custom_call_target="__cublas$lt$groupedMatmul")");
-  EXPECT_TRUE(RunAndCompare(hlo_text, ErrorSpec{1e-2, 1e-2}));
-}
-
-TEST_F(GroupedGemmRewriteTest,
-       CheckCustomCallTargetGroupedGemmRaggedDimInContractingDimWithBatchDim) {
-  if (SkipGpuBlasLtTest()) {
-    GTEST_SKIP() << "BlasLt is not supported on this GPU architecture";
-  }
-  const char* hlo_text = R"(
-HloModule GroupedGemm
-
-ENTRY AddRaggedDotsFunc {
-    p0 = bf16[3,64,9]{2,1,0} parameter(0)
-    p1 = bf16[3,9,8]{2,1,0} parameter(1)
-    p2 = s64[3,2] constant({{4, 5}, {4, 5}, {4, 5}})
-    ROOT ragged-dot = bf16[2,3,64,8]{3,2,1,0} ragged-dot(p0, p1, p2),
-                      lhs_contracting_dims={2}, rhs_contracting_dims={1},
-                      lhs_ragged_dims={2}, lhs_batch_dims={0}, rhs_batch_dims={0}
-}
-)";
-  MatchOptimizedHlo(hlo_text,
-                    R"(
-                    ; CHECK: custom_call_target="__cublas$lt$groupedMatmul", 
-                    ; CHECK-SAME: backend_config={"operation_queue_id":"0","wait_on_operation_queues":[],
-                    ; CHECK-SAME: "force_earliest_schedule":false,"reification_cost":[],
-                    ; CHECK-SAME: "device_type":"DEVICE_TYPE_INVALID",
-                    ; CHECK-SAME: "grouped_gemm_backend_config":{
-                    ; CHECK-SAME: "gemm_backend_config":{"alpha_real":1,"beta":0,
-                    ; CHECK-SAME: "dot_dimension_numbers":{"lhs_contracting_dimensions":["2"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["1"],"lhs_batch_dimensions":["0"],
-                    ; CHECK-SAME: "rhs_batch_dimensions":["0"]},"alpha_imag":0,"epilogue":"DEFAULT",
-                    ; CHECK-SAME: "grad_x":false,"grad_y":false,
-                    ; CHECK-SAME: "damax_output":false,"autotune_workspace_size":"0"},
-                    ; CHECK-SAME: "ragged_dot_dimension_numbers":{"dot_dimension_numbers":{
-                    ; CHECK-SAME: "lhs_contracting_dimensions":["2"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["1"],
-                    ; CHECK-SAME: "lhs_batch_dimensions":["0"],
-                    ; CHECK-SAME: "rhs_batch_dimensions":["0"]},
-                    ; CHECK-SAME: "lhs_ragged_dimensions":["2"],
-                    ; CHECK-SAME: "rhs_group_dimensions":[]}}})");
-  EXPECT_TRUE(RunAndCompare(hlo_text, ErrorSpec{1e-4, 1e-5}));
-}
-
-TEST_F(GroupedGemmRewriteTest,
-       CheckCustomCallTargetGroupedGemmRaggedDimInBatchDim) {
-  if (SkipGpuBlasLtTest()) {
-    GTEST_SKIP() << "BlasLt is not supported on this GPU architecture";
-  }
-  const char* hlo_text = R"(
-HloModule GroupedGemm
-
-ENTRY AddRaggedDotsFunc {
-    p0 = bf16[5,16,9]{2,1,0} parameter(0)
-    p1 = bf16[5,9,8]{2,1,0} parameter(1)
-    p2 = s64[2] constant({3, 2})
-    ROOT ragged-dot = bf16[5,16,8]{2,1,0} ragged-dot(p0, p1, p2),
-                      lhs_contracting_dims={2}, rhs_contracting_dims={1},
-                      lhs_ragged_dims={0}, lhs_batch_dims={0}, rhs_batch_dims={0}
-}
-)";
-  MatchOptimizedHlo(hlo_text,
-                    R"(
-                    ; CHECK: custom_call_target="__cublas$lt$groupedMatmul", 
-                    ; CHECK-SAME: backend_config={"operation_queue_id":"0","wait_on_operation_queues":[],
-                    ; CHECK-SAME: "force_earliest_schedule":false,"reification_cost":[],
-                    ; CHECK-SAME: "device_type":"DEVICE_TYPE_INVALID",
-                    ; CHECK-SAME: "grouped_gemm_backend_config":{
-                    ; CHECK-SAME: "gemm_backend_config":{"alpha_real":1,"beta":0,
-                    ; CHECK-SAME: "dot_dimension_numbers":{"lhs_contracting_dimensions":["2"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["1"],"lhs_batch_dimensions":["0"],
-                    ; CHECK-SAME: "rhs_batch_dimensions":["0"]},"alpha_imag":0,"epilogue":"DEFAULT",
-                    ; CHECK-SAME: "grad_x":false,"grad_y":false,
-                    ; CHECK-SAME: "damax_output":false,"autotune_workspace_size":"0"},
-                    ; CHECK-SAME: "ragged_dot_dimension_numbers":{"dot_dimension_numbers":{
-                    ; CHECK-SAME: "lhs_contracting_dimensions":["2"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["1"],
-                    ; CHECK-SAME: "lhs_batch_dimensions":["0"],
-                    ; CHECK-SAME: "rhs_batch_dimensions":["0"]},
-                    ; CHECK-SAME: "lhs_ragged_dimensions":["0"],
-                    ; CHECK-SAME: "rhs_group_dimensions":[]}}})");
-  EXPECT_TRUE(RunAndCompare(hlo_text, ErrorSpec{1e-3, 1e-3}));
-}
-
-TEST_F(GroupedGemmRewriteTest,
-       CheckCustomCallTargetGroupedGemmRaggedDimInBatchDimMultipleGroups) {
-  if (SkipGpuBlasLtTest()) {
-    GTEST_SKIP() << "BlasLt is not supported on this GPU architecture";
-  }
-  const char* hlo_text = R"(
-HloModule GroupedGemm
-
-ENTRY AddRaggedDotsFunc {
-    p0 = bf16[16,64,9]{2,1,0} parameter(0)
-    p1 = bf16[16,9,8]{2,1,0} parameter(1)
-    p2 = s64[4] constant({4, 2, 6, 4})
-    ROOT ragged-dot = bf16[16,64,8]{2,1,0} ragged-dot(p0, p1, p2),
-                      lhs_contracting_dims={2}, rhs_contracting_dims={1},
-                      lhs_ragged_dims={0}, lhs_batch_dims={0}, rhs_batch_dims={0}
-}
-)";
-  MatchOptimizedHlo(hlo_text,
-                    R"(
-                    ; CHECK: custom_call_target="__cublas$lt$groupedMatmul", 
-                    ; CHECK-SAME: backend_config={"operation_queue_id":"0","wait_on_operation_queues":[],
-                    ; CHECK-SAME: "force_earliest_schedule":false,"reification_cost":[],
-                    ; CHECK-SAME: "device_type":"DEVICE_TYPE_INVALID",
-                    ; CHECK-SAME: "grouped_gemm_backend_config":{
-                    ; CHECK-SAME: "gemm_backend_config":{"alpha_real":1,"beta":0,
-                    ; CHECK-SAME: "dot_dimension_numbers":{"lhs_contracting_dimensions":["2"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["1"],"lhs_batch_dimensions":["0"],
-                    ; CHECK-SAME: "rhs_batch_dimensions":["0"]},"alpha_imag":0,"epilogue":"DEFAULT",
-                    ; CHECK-SAME: "grad_x":false,"grad_y":false,
-                    ; CHECK-SAME: "damax_output":false,"autotune_workspace_size":"0"},
-                    ; CHECK-SAME: "ragged_dot_dimension_numbers":{"dot_dimension_numbers":{
-                    ; CHECK-SAME: "lhs_contracting_dimensions":["2"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["1"],
-                    ; CHECK-SAME: "lhs_batch_dimensions":["0"],
-                    ; CHECK-SAME: "rhs_batch_dimensions":["0"]},
-                    ; CHECK-SAME: "lhs_ragged_dimensions":["0"],
-                    ; CHECK-SAME: "rhs_group_dimensions":[]}}})");
-  EXPECT_TRUE(RunAndCompare(hlo_text, ErrorSpec{1e-4, 1e-5}));
-}
-
-TEST_F(
-    GroupedGemmRewriteTest,
-    CheckCustomCallTargetGroupedGemmRaggedDimInBatchDimMultipleGroupsOutputColumnMajor) {
-  if (SkipGpuBlasLtTest()) {
-    GTEST_SKIP() << "BlasLt is not supported on this GPU architecture";
-  }
-  const char* hlo_text = R"(
-HloModule GroupedGemm
-
-ENTRY AddRaggedDotsFunc {
-    p0 = bf16[16,64,9]{2,1,0} parameter(0)
-    p1 = bf16[16,9,8]{2,1,0} parameter(1)
-    p2 = s64[4] constant({4, 2, 6, 4})
-    ROOT ragged-dot = bf16[16,64,8]{1,2,0} ragged-dot(p0, p1, p2),
-                      lhs_contracting_dims={2}, rhs_contracting_dims={1},
-                      lhs_ragged_dims={0}, lhs_batch_dims={0}, rhs_batch_dims={0}
-}
-)";
-  MatchOptimizedHlo(hlo_text,
-                    R"(
-                    ; CHECK: custom_call_target="__cublas$lt$groupedMatmul", 
-                    ; CHECK-SAME: backend_config={"operation_queue_id":"0","wait_on_operation_queues":[],
-                    ; CHECK-SAME: "force_earliest_schedule":false,"reification_cost":[],
-                    ; CHECK-SAME: "device_type":"DEVICE_TYPE_INVALID",
-                    ; CHECK-SAME: "grouped_gemm_backend_config":{
-                    ; CHECK-SAME: "gemm_backend_config":{"alpha_real":1,"beta":0,
-                    ; CHECK-SAME: "dot_dimension_numbers":{"lhs_contracting_dimensions":["2"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["1"],"lhs_batch_dimensions":["0"],
-                    ; CHECK-SAME: "rhs_batch_dimensions":["0"]},"alpha_imag":0,"epilogue":"DEFAULT",
-                    ; CHECK-SAME: "grad_x":false,"grad_y":false,
-                    ; CHECK-SAME: "damax_output":false,"autotune_workspace_size":"0"},
-                    ; CHECK-SAME: "ragged_dot_dimension_numbers":{"dot_dimension_numbers":{
-                    ; CHECK-SAME: "lhs_contracting_dimensions":["2"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["1"],
-                    ; CHECK-SAME: "lhs_batch_dimensions":["0"],
-                    ; CHECK-SAME: "rhs_batch_dimensions":["0"]},
-                    ; CHECK-SAME: "lhs_ragged_dimensions":["0"],
-                    ; CHECK-SAME: "rhs_group_dimensions":[]}}})");
-  EXPECT_TRUE(RunAndCompare(hlo_text, ErrorSpec{1e-4, 1e-5}));
-}
-
-TEST_F(GroupedGemmRewriteTest,
-       CheckCustomCallTargetGroupedGemmRaggedDimInBatchDimLargeGroupCount600) {
-  if (SkipGpuBlasLtTest()) {
-    GTEST_SKIP() << "BlasLt is not supported on this GPU architecture";
-  }
-  // Test with 600 groups (> 256) for RaggedDimInBatchDim
-  // Sum of group sizes: 600 * 4 = 2400 (matches p0 first dim for batch)
-  const char* hlo_text = R"(
-HloModule GroupedGemmRaggedBatchLarge600
-
-ENTRY AddRaggedDotsFunc {
-    p0 = bf16[2400,64,9]{2,1,0} parameter(0)
-    p1 = bf16[2400,9,8]{2,1,0} parameter(1)
-    scalar = s64[] constant(4)
-    p2 = s64[600] broadcast(scalar), dimensions={}
-    ROOT ragged-dot = bf16[2400,64,8]{2,1,0} ragged-dot(p0, p1, p2),
-                      lhs_contracting_dims={2}, rhs_contracting_dims={1},
-                      lhs_ragged_dims={0}, lhs_batch_dims={0}, rhs_batch_dims={0}
-}
-)";
-  MatchOptimizedHlo(hlo_text,
-                    R"(
-                    ; CHECK: custom_call_target="__cublas$lt$groupedMatmul")");
-  EXPECT_TRUE(RunAndCompare(hlo_text, ErrorSpec{1e-2, 1e-2}));
-}
-
-TEST_F(
-    GroupedGemmRewriteTest,
-    CheckCustomCallTargetGroupedGemmRaggedInNonContractingGroupDimNoOuterDim) {
-  if (SkipGpuBlasLtTest()) {
-    GTEST_SKIP() << "BlasLt is not supported on this GPU architecture";
-  }
-  const char* hlo_text = R"(
-HloModule GroupedGemm
-
-ENTRY AddRaggedDotsFunc {
-    p0 = bf16[128,64]{1,0} parameter(0)
-    p1 = bf16[8,64,32]{2,0,1} parameter(1)
-    p2 = s32[8] constant({16, 16, 16, 16, 16, 16, 16, 16})
-    ROOT ragged-dot = bf16[128,32]{1,0} ragged-dot(p0, p1, p2),
-                      lhs_contracting_dims={1}, rhs_contracting_dims={1},
-                      lhs_ragged_dims={0}, rhs_group_dims={0}
-}
-)";
-  MatchOptimizedHlo(hlo_text,
-                    R"(
-                    ; CHECK: custom_call_target="__cublas$lt$groupedMatmul", 
-                    ; CHECK-SAME: backend_config={"operation_queue_id":"0","wait_on_operation_queues":[],
-                    ; CHECK-SAME: "force_earliest_schedule":false,"reification_cost":[],
-                    ; CHECK-SAME: "device_type":"DEVICE_TYPE_INVALID",
-                    ; CHECK-SAME: "grouped_gemm_backend_config":{
-                    ; CHECK-SAME: "gemm_backend_config":{"alpha_real":1,"beta":0,
-                    ; CHECK-SAME: "dot_dimension_numbers":{"lhs_contracting_dimensions":["1"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["1"],"lhs_batch_dimensions":[],
-                    ; CHECK-SAME: "rhs_batch_dimensions":[]},"alpha_imag":0,"epilogue":"DEFAULT",
-                    ; CHECK-SAME: "grad_x":false,"grad_y":false,
-                    ; CHECK-SAME: "damax_output":false,"autotune_workspace_size":"0"},
-                    ; CHECK-SAME: "ragged_dot_dimension_numbers":{"dot_dimension_numbers":{
-                    ; CHECK-SAME: "lhs_contracting_dimensions":["1"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["1"],
-                    ; CHECK-SAME: "lhs_batch_dimensions":[],
-                    ; CHECK-SAME: "rhs_batch_dimensions":[]},
-                    ; CHECK-SAME: "lhs_ragged_dimensions":["0"],
-                    ; CHECK-SAME: "rhs_group_dimensions":["0"]}}})");
-  EXPECT_TRUE(RunAndCompare(hlo_text, ErrorSpec{1e-3, 1e-3}));
-}
-
-TEST_F(GroupedGemmRewriteTest,
-       CheckCustomCallTargetGroupedGemmRaggedInContractingGroupDimNoOuterDim) {
-  if (SkipGpuBlasLtTest()) {
-    GTEST_SKIP() << "BlasLt is not supported on this GPU architecture";
-  }
-  const char* hlo_text = R"(
-HloModule GroupedGemm
-
-ENTRY AddRaggedDotsFunc {
-    p0 = bf16[64,1024]{1,0} parameter(0)
-    p1 = bf16[1024,256]{1,0} parameter(1)
-    p2 = s32[8] constant({128, 128, 128, 128, 128, 128, 128, 128})
-    ROOT ragged-dot = bf16[8,64,256]{2,0,1} ragged-dot(p0, p1, p2),
-                      lhs_contracting_dims={1}, rhs_contracting_dims={0},
-                      lhs_ragged_dims={1}
-}
-)";
-  MatchOptimizedHlo(hlo_text,
-                    R"(
-                    ; CHECK: custom_call_target="__cublas$lt$groupedMatmul", 
-                    ; CHECK-SAME: backend_config={"operation_queue_id":"0","wait_on_operation_queues":[],
-                    ; CHECK-SAME: "force_earliest_schedule":false,"reification_cost":[],
-                    ; CHECK-SAME: "device_type":"DEVICE_TYPE_INVALID",
-                    ; CHECK-SAME: "grouped_gemm_backend_config":{
-                    ; CHECK-SAME: "gemm_backend_config":{"alpha_real":1,"beta":0,
-                    ; CHECK-SAME: "dot_dimension_numbers":{"lhs_contracting_dimensions":["1"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["0"],"lhs_batch_dimensions":[],
-                    ; CHECK-SAME: "rhs_batch_dimensions":[]},"alpha_imag":0,"epilogue":"DEFAULT",
-                    ; CHECK-SAME: "grad_x":false,"grad_y":false,
-                    ; CHECK-SAME: "damax_output":false,"autotune_workspace_size":"0"},
-                    ; CHECK-SAME: "ragged_dot_dimension_numbers":{"dot_dimension_numbers":{
-                    ; CHECK-SAME: "lhs_contracting_dimensions":["1"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["0"],
-                    ; CHECK-SAME: "lhs_batch_dimensions":[],
-                    ; CHECK-SAME: "rhs_batch_dimensions":[]},
-                    ; CHECK-SAME: "lhs_ragged_dimensions":["1"],
-                    ; CHECK-SAME: "rhs_group_dimensions":[]}}})");
-  EXPECT_TRUE(RunAndCompare(hlo_text, ErrorSpec{1e-2, 1e-2}));
-}
-
-TEST_F(GroupedGemmRewriteTest,
-       CheckCustomCallTargetGroupedGemmRaggedInContractingTranspose) {
-  if (SkipGpuBlasLtTest()) {
-    GTEST_SKIP() << "BlasLt is not supported on this GPU architecture";
-  }
-  const char* hlo_text = R"(
-HloModule GroupedGemm
-
-ENTRY AddRaggedDotsFunc {
-    p0 = bf16[1024,64]{1,0} parameter(0)
-    p1 = bf16[1024,256]{1,0} parameter(1)
-    p2 = s32[8] constant({128, 128, 128, 128, 128, 128, 128, 128})
-    ROOT ragged-dot = bf16[8,64,256]{2,0,1} ragged-dot(p0, p1, p2),
-                      lhs_contracting_dims={0}, rhs_contracting_dims={0},
-                      lhs_ragged_dims={0}
-}
-)";
-  MatchOptimizedHlo(hlo_text,
-                    R"(
-                    ; CHECK: custom_call_target="__cublas$lt$groupedMatmul", 
-                    ; CHECK-SAME: backend_config={"operation_queue_id":"0","wait_on_operation_queues":[],
-                    ; CHECK-SAME: "force_earliest_schedule":false,"reification_cost":[],
-                    ; CHECK-SAME: "device_type":"DEVICE_TYPE_INVALID",
-                    ; CHECK-SAME: "grouped_gemm_backend_config":{
-                    ; CHECK-SAME: "gemm_backend_config":{"alpha_real":1,"beta":0,
-                    ; CHECK-SAME: "dot_dimension_numbers":{"lhs_contracting_dimensions":["0"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["0"],"lhs_batch_dimensions":[],
-                    ; CHECK-SAME: "rhs_batch_dimensions":[]},"alpha_imag":0,"epilogue":"DEFAULT",
-                    ; CHECK-SAME: "grad_x":false,"grad_y":false,
-                    ; CHECK-SAME: "damax_output":false,"autotune_workspace_size":"0"},
-                    ; CHECK-SAME: "ragged_dot_dimension_numbers":{"dot_dimension_numbers":{
-                    ; CHECK-SAME: "lhs_contracting_dimensions":["0"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["0"],
-                    ; CHECK-SAME: "lhs_batch_dimensions":[],
-                    ; CHECK-SAME: "rhs_batch_dimensions":[]},
-                    ; CHECK-SAME: "lhs_ragged_dimensions":["0"],
-                    ; CHECK-SAME: "rhs_group_dimensions":[]}}})");
-  EXPECT_TRUE(RunAndCompare(hlo_text, ErrorSpec{1e-2, 1e-2}));
-}
-
-TEST_F(GroupedGemmRewriteTest,
-       CheckCustomCallTargetGroupedGemmRaggedNonContractingTranspose) {
-  if (SkipGpuBlasLtTest()) {
-    GTEST_SKIP() << "BlasLt is not supported on this GPU architecture";
-  }
-  const char* hlo_text = R"(
-HloModule GroupedGemm
-
-ENTRY AddRaggedDotsFunc {
-    p0 = bf16[9,64]{1,0} parameter(0)
-    p1 = bf16[4,9,8]{2,1,0} parameter(1)
-    p2 = s64[4] constant({16, 8, 24, 16})
-    ROOT ragged-dot = bf16[64,8]{1,0} ragged-dot(p0, p1, p2),
-                      lhs_contracting_dims={0}, rhs_contracting_dims={1},
-                      lhs_ragged_dims={1}, rhs_group_dims={0}
-}
-)";
-  MatchOptimizedHlo(hlo_text,
-                    R"(
-                    ; CHECK: custom_call_target="__cublas$lt$groupedMatmul", 
-                    ; CHECK-SAME: backend_config={"operation_queue_id":"0","wait_on_operation_queues":[],
-                    ; CHECK-SAME: "force_earliest_schedule":false,"reification_cost":[],
-                    ; CHECK-SAME: "device_type":"DEVICE_TYPE_INVALID",
-                    ; CHECK-SAME: "grouped_gemm_backend_config":{
-                    ; CHECK-SAME: "gemm_backend_config":{"alpha_real":1,"beta":0,
-                    ; CHECK-SAME: "dot_dimension_numbers":{"lhs_contracting_dimensions":["0"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["1"],"lhs_batch_dimensions":[],
-                    ; CHECK-SAME: "rhs_batch_dimensions":[]},"alpha_imag":0,"epilogue":"DEFAULT",
-                    ; CHECK-SAME: "grad_x":false,"grad_y":false,
-                    ; CHECK-SAME: "damax_output":false,"autotune_workspace_size":"0"},
-                    ; CHECK-SAME: "ragged_dot_dimension_numbers":{"dot_dimension_numbers":{
-                    ; CHECK-SAME: "lhs_contracting_dimensions":["0"],
-                    ; CHECK-SAME: "rhs_contracting_dimensions":["1"],
-                    ; CHECK-SAME: "lhs_batch_dimensions":[],
-                    ; CHECK-SAME: "rhs_batch_dimensions":[]},
-                    ; CHECK-SAME: "lhs_ragged_dimensions":["1"],
-                    ; CHECK-SAME: "rhs_group_dimensions":["0"]}}})");
-  EXPECT_TRUE(RunAndCompare(hlo_text, ErrorSpec{1e-4, 1e-5}));
-}
-
 TEST_F(GemmRewriteTest, CheckCustomCallTarget) {
   if (SkipGpuBlasLtTest()) {
     GTEST_SKIP() << "BlasLt is not supported on this GPU architecture";
